+++ lib/src/client.dart
@@ -1,12 +1,14 @@
 library imgur.client;
 
 import 'dart:io';
-
 import 'package:http/http.dart' as http;
 import 'package:imgur/imgur.dart';
+import 'package:json_annotation/json_annotation.dart';
 import 'dart:convert';
 
 import 'common.dart';
+
+part 'client.g.dart';
 
 part 'client/services/account.dart';
 part 'client/services/service.dart';