+++ lib/src/models/gallery_album_image.dart
@@ -64,7 +64,19 @@ class GalleryAlbumImage implements BaseModel {
   bool isAlbum;
 
   /// The current user's vote on the album. null if not signed in or if the user hasn't voted on it
-  String vote;
+  @JsonKey(fromJson: stringToVote, toJson: voteToString)
+  VoteType vote;
+
+  @JsonKey(name: 'favorite_count')
+  int favoriteCount;
+
+  /// A list of tags associated with the image.
+  @JsonKey(toJson: baseModelListToJson)
+  List<Tag> tags;
+
+  /// If the video has sound or not.
+  @JsonKey(nullable: true, name: 'has_sound')
+  bool hasSound;
 
   /// Indicates if the current user favorited the album. Defaults to false if not signed in
   bool favorite;
@@ -88,7 +100,7 @@ class GalleryAlbumImage implements BaseModel {
   int imagesCount;
 
   /// An array of all the images in the album (only available when requesting the direct album)
-  @JsonKey(nullable: true, toJson: baseModelListToJson)
+  @JsonKey(toJson: baseModelListToJson)
   List<Image> images;
 
   @JsonKey(name: 'in_most_viral')
@@ -145,6 +157,9 @@ class GalleryAlbumImage implements BaseModel {
     this.privacy,
     this.layout,
     this.views,
+    this.favoriteCount,
+    this.tags,
+    this.hasSound,
     this.link,
     this.ups,
     this.downs,