+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.0.2
+
+- Change GalleryImage accountUrl field type from int to String
+- Add error handling for null input in baseModelListToJson method
+- Add some unit tests
+- Add getVotes and vote methods to image and album services
+- Add favoriteCount, tags and hasSound to Image and Album models
+- Add headers getter to Authentication class
+
 ## 1.0.1
 
 - Minor fixes based on pub.dev suggestions